added and cleaned up whoop tool calls. moved answer_with_tools to a dedicated module
